Adding authentication

+++ backend/routes/posts.js
@@ -1,7 +1,8 @@
 const express = require('express');
-const Post = require('../models/post');
 const multer = require('multer');
 
+const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
 
@@ -12,6 +13,8 @@ const MIME_TYPE_MAP = {
 }
 var fetchedPosts;
 
+
+//creating a multer for image
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
     const isValid = MIME_TYPE_MAP[file.mimetype];
@@ -28,12 +31,14 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post('', multer({storage: storage}).single("image"), (req, res, next) => {
+//post a Post
+router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   const url = req.protocol+ '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()
     .then(result => {
@@ -47,6 +52,7 @@ router.post('', multer({storage: storage}).single("image"), (req, res, next) => 
     })
 });
 
+//Get all post
 router.get('', (req, res, next) => {
   const pageSize = +req.query.pageSize;
   const currentPage = +req.query.page;
@@ -70,13 +76,20 @@ router.get('', (req, res, next) => {
     })
 })
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+
+//Delete a post
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if(result.n != 0)
     res.status(200).json({ message: "Post deleted!" });
+    else
+      res.status(401).json({message: 'Not Autherized'})
   });
 });
 
-router.put("/:id", multer({storage: storage}).single("image"), (req, res, next) => {
+
+//updtae  particular post
+router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
 
   let imagePath = req.body.imagePath;
   if(req.file) {
@@ -88,14 +101,20 @@ router.put("/:id", multer({storage: storage}).single("image"), (req, res, next) 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if(result.nModified != 0)
+      res.status(200).json({ message: "Update successful!" });
+    else
+      res.status(401).json({message: 'Not Autherized'})
   });
 });
 
+
+//get a particular post
 router.get("/:id", (req, res, next) => {
   Post.findById(req.params.id).then(post => {
     if(post) {
